pyauto-ecm-scrape-lazada-windows-master/requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
MouseInfo==0.1.3
mss==9.0.1
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.25.0
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pygsheets==2.0.6
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
rubicon-objc==0.4.7
six==1.16.0
soupsieve==2.5
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
wget==3.2
xlrd==2.0.1
